chore: fix ci

+++ src/modules/GroupChannel/components/RemoveMessageModal.tsx
@@ -15,26 +15,28 @@ export interface RemoveMessageProps {
 
 const RemoveMessage: React.FC<RemoveMessageProps> = (props: RemoveMessageProps) => {
   const {
-    onSubmit = () => { /* noop */ },
+    onSubmit = () => {
+      /* noop */
+    },
     onCancel,
     message,
   } = props;
   const { stringSet } = useContext(LocalizationContext);
-  const {
-    deleteMessage,
-  } = useGroupChannelContext();
+  const { deleteMessage } = useGroupChannelContext();
   return (
     <Modal
       type={ButtonTypes.DANGER}
       disabled={message?.threadInfo?.replyCount > 0}
       onCancel={onCancel}
       onSubmit={() => {
-        deleteMessage(message).then(() => {
-          // For other status such as PENDING, SCHEDULED, and CANCELED,
-          // invalid parameters error is thrown so nothing happens.
-          onSubmit();
-          onCancel();
-        });
+        if (message.isUserMessage() || message.isFileMessage() || message.isMultipleFilesMessage()) {
+          deleteMessage(message).then(() => {
+            // For other status such as PENDING, SCHEDULED, and CANCELED,
+            // invalid parameters error is thrown so nothing happens.
+            onSubmit();
+            onCancel();
+          });
+        }
       }}
       submitText={stringSet.MESSAGE_MENU__DELETE}
       titleText={getModalDeleteMessageTitle(stringSet, message)}
